% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_patterns.R
\name{color_patterns_get}
\alias{color_patterns_get}
\title{Set up col, pattern and group for plotting.}
\usage{
color_patterns_get(scan1ggdata, col, palette = NULL, shape)
}
\arguments{
\item{scan1ggdata}{data frame to be used for plotting}

\item{col}{Color for \code{color} column in \code{scan1ggdata}}

\item{palette}{for colors (default \code{NULL} uses \code{"Dark2"} from \code{RColorBrewer} package)}
}
\description{
Set up col, pattern and group for plotting.
}
