% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_patterns.R
\name{color_patterns_pheno}
\alias{color_patterns_pheno}
\title{Set up pheno and pattern, and chr_pheno, for colors.}
\usage{
color_patterns_pheno(scan1ggdata, lod, pattern, col, patterns, facet)
}
\description{
Set up pheno and pattern, and chr_pheno, for colors.
}
