% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_genes_internal.R
\name{ggplot_genes_internal}
\alias{ggplot_genes_internal}
\title{GGPlot internal routine for ggplot_genes}
\usage{
ggplot_genes_internal(
  start,
  end,
  strand,
  rect_top,
  rect_bottom,
  colors,
  space,
  y,
  dir_symbol,
  name,
  xlim,
  xlab = "Position (Mbp)",
  ylab = "",
  bgcolor = "gray92",
  xat = NULL,
  legend.position = "none",
  vlines = NULL,
  ...
)
}
\arguments{
\item{start, end, strand, rect_top, rect_bottom, colors, space, y, dir_symbol, name, xlim}{usual parameters}

\item{legend.position, vlines, xlab, ylab, bgcolor, xat}{hidden parameters}

\item{...}{Additional graphics parameters.}
}
\value{
object of class \code{\link[ggplot2]{ggplot}}.
}
\description{
Plot genes at positions
}
\keyword{internal}
