% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_listof_scan1coef.R
\name{ggplot_listof_scan1coef}
\alias{ggplot_listof_scan1coef}
\alias{autoplot.listof_scan1coef}
\title{Plot of object of class listof_scan1coeff}
\usage{
ggplot_listof_scan1coef(x, map, columns = NULL, col = NULL,
  scan1_output = NULL, facet = "pattern", ...)

\method{autoplot}{listof_scan1coef}(x, ...)
}
\arguments{
\item{x}{object of class \code{listof_scan1coeff}}

\item{map}{A list of vectors of marker positions, as produced by
\code{\link[qtl2]{insert_pseudomarkers}}.}

\item{columns}{Vector of columns to plot}

\item{col}{Vector of colors, same length as \code{columns}. If
NULL, some default choices are made.}

\item{scan1_output}{If provided, we make a two-panel plot with
coefficients on top and LOD scores below. Should have just one LOD
score column; if multiple, only the first is used.}

\item{facet}{Plot facets if multiple phenotypes and group provided (default = \code{"pattern"}).}

\item{...}{arguments for \code{\link{ggplot_coef}}}

\item{pattern}{Use phenotype names as pattern.}
}
\description{
Plot object of class \code{listof_scan1coeff}, which is a list of objects of class \code{scan1coef}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}
