% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scan1_internal.R
\name{ggplot_scan1_internal}
\alias{ggplot_scan1_internal}
\title{Plot a genome scan}
\usage{
ggplot_scan1_internal(
  map,
  lod,
  gap = 25,
  col = NULL,
  shape = NULL,
  pattern = NULL,
  facet = NULL,
  patterns = c("none", "all", "hilit"),
  ...
)
}
\arguments{
\item{map}{Map of pseudomarker locations.}

\item{lod}{Matrix of lod (or other) values.}

\item{gap}{Gap between chromosomes.}

\item{col}{Colors for points or lines, with labels.}

\item{shape}{Shapes for points.}

\item{pattern}{Use to group values for plotting (default = \code{NULL}); typically provided by \code{\link{plot_snpasso}} internal routine.}

\item{facet}{Plot facets if multiple phenotypes and group provided (default = \code{NULL}).}

\item{patterns}{Connect SDP patterns: one of \code{c("none","all","hilit")}.}

\item{...}{Additional graphics parameters.}
}
\description{
Plot LOD curves for a genome scan
}
