% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scan_pattern.R
\name{ggplot_scan_pattern}
\alias{ggplot_scan_pattern}
\alias{autoplot.scan_pattern}
\title{Plot scan pattern}
\usage{
ggplot_scan_pattern(x, map, plot_type = c("lod", "coef", "coef_and_lod"),
  patterns = x$patterns$founders, columns = 1:3, min_lod = 3,
  lodcolumn = seq_along(patterns), facet = "pheno", ...)

\method{autoplot}{scan_pattern}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{scan_pattern}}}

\item{map}{A list of vectors of marker positions, as produced by
\code{\link[qtl2]{insert_pseudomarkers}}.}

\item{plot_type}{type of plot from \code{c("lod","coef")}}

\item{patterns}{allele patterns to plot (default all)}

\item{columns}{columns used for coef plot}

\item{min_lod}{minimum LOD peak for contrast to be retained}

\item{lodcolumn}{columns used for scan1 plot (default all \code{patterns})}

\item{facet}{Plot facets if multiple phenotypes and patterns provided (default = \code{"pheno"}).}

\item{...}{additional parameters}
}
\description{
Plot scan pattern
}
