% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listof__scan1coeff.R
\name{listof_scan1coef}
\alias{listof_scan1coef}
\alias{plot.listof_scan1coef}
\alias{summary.listof_scan1coef}
\alias{summary.scan1coef}
\title{List of scan1coef objects}
\usage{
listof_scan1coef(probs, phe, K = NULL, covar = NULL)

\method{plot}{listof_scan1coef}(x, columns = NULL, col = NULL,
  facet = "pattern", ...)

\method{summary}{listof_scan1coef}(object, scan1_object,
  coef_names = dimnames(object[[1]]$coef)[[2]], ...)

\method{summary}{scan1coef}(object, scan1_object, ...)
}
\arguments{
\item{probs}{genotype probabilities object for one chromosome from \code{\link[qtl2geno]{calc_genoprob}}}

\item{phe}{data frame of phenotypes}

\item{K}{list of length 1 with kinship matrix}

\item{covar}{matrix of covariates}

\item{x}{object of class \code{listof_scan1coeff}}

\item{facet}{Plot facets if multiple phenotypes and group provided (default = \code{"pattern"}).}

\item{...}{arguments for \code{\link[qtl2scan]{plot_coef}}}

\item{object}{object of class \code{listof_scan1coeff}}

\item{scan1_object}{object from \code{scan1}}

\item{coef_names}{names of effect coefficients (default is all coefficient names)}

\item{...}{arguments for \code{\link[qtl2scan]{plot_coef}}}

\item{object}{object of class \code{listof_scan1coeff}}

\item{scan1_object}{object from \code{scan1}}

\item{...}{arguments for \code{\link[qtl2scan]{plot_coef}}}
}
\value{
object of class \code{listof_scan1coeff}
}
\description{
Create a list of scan1coef objects using \code{\link[qtl2scan]{scan1coef}}.

Plot object of class \code{listof_scan1coeff}, which is a list of objects of class \code{scan1coef}.

Summary of object of class \code{\link{listof_scan1coeff}}, which is a list of objects of class \code{scan1coef}.

Summary of object of class \code{\link{listof_scan1coeff}}, which is a list of objects of class \code{scan1coef}.
}
\examples{
\dontrun{listof_scan1coef(probs, phe)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}

Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}
\keyword{utilities}
