% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pxg.R
\name{mean_pxg}
\alias{mean_pxg}
\title{Mean phenotype by genotype}
\usage{
mean_pxg(geno, pheno, dataframe = NULL)
}
\arguments{
\item{geno}{Vector of genotypes, as produced by
\code{\link[qtl2]{maxmarg}} with specific \code{chr} and
\code{pos}.}

\item{pheno}{Vector of phenotypes.}

\item{dataframe}{Supplied data frame, or constructed from \code{geno} and \code{pheno} if \code{NULL}.}
}
\description{
Mean phenotype by genotype
}
