% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.listof_scan1coef.R
\name{plot.listof_scan1coef}
\alias{plot.listof_scan1coef}
\title{Plot of object of class listof_scan1coeff}
\usage{
\method{plot}{listof_scan1coef}(x, columns = NULL, col = NULL,
  facet = "pattern", ...)
}
\arguments{
\item{x}{object of class \code{listof_scan1coeff}}

\item{facet}{Plot facets if multiple phenotypes and group provided (default = \code{"pattern"}).}

\item{...}{arguments for \code{\link[qtl2scan]{plot_coef}}}
}
\description{
Plot object of class \code{listof_scan1coeff}, which is a list of objects of class \code{scan1coef}.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}
