% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allele1.R
\name{plot_allele1}
\alias{plot_allele1}
\title{Plot allele1 object}
\usage{
plot_allele1(x, scan1_object = NULL, map = NULL, pos = NULL,
  trim = TRUE, frame = FALSE, ...)
}
\arguments{
\item{x}{Object of class \code{\link[qtl2pattern]{allele1}}.}

\item{scan1_object}{Optional object of class \code{\link[qtl2scan]{scan1}} to find peak.}

\item{map}{Genome map (required if \code{scan1_object} present)}

\item{pos}{Genome position in Mbp (supercedes \code{scan1_object})}

\item{trim}{If \code{TRUE}, trim extreme alleles.}

\item{frame}{If \code{TRUE}, enable frames for \code{\link[plotly]{ggplotly}}}

\item{...}{Other parameters ignored.}
}
\description{
Plot alleles for haplotype, diplotype and top patterns and genome position.
}
