% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_feature_tbl.R
\name{plot_feature_tbl}
\alias{plot_feature_tbl}
\alias{autoplot.feature_tbl}
\alias{plot.feature_tbl}
\title{Staggered plot of genes in interval}
\usage{
plot_feature_tbl(x, rect_col = "grey70", strand_col = c(`-` = "#1b9e77", `+`
  = "#d95f02"), type_col = c(gene = "black", pseudogene = "#1b9e77", other =
  "#d95f02"), text_size = 3, xlim = NULL, snp_pos = top_snps_tbl$pos_Mbp,
  snp_lod = top_snps_tbl$lod, top_snps_tbl = NULL, snp_col = "grey70",
  extend = 5000, ...)

\method{autoplot}{feature_tbl}(x, ...)

\method{plot}{feature_tbl}(x, ...)
}
\arguments{
\item{x}{tbl of gene information from \code{\link{get_mgi_features}}}

\item{rect_col}{fill color of rectangle (default "grey70")}

\item{strand_col}{edge color of rectangle by strand from \code{x} (default -="blue", +="red"; none if NULL)}

\item{type_col}{color of type from \code{x} (default "black" for gene, "blue" for pseudogene; none if NULL)}

\item{text_size}{size of text (default 3)}

\item{xlim}{horizontal axis limits (default is range of features)}

\item{snp_pos}{position of SNPs in bp if used (default NULL)}

\item{snp_lod}{LOD of SNPs (for color plotting)}

\item{top_snps_tbl}{table from \code{\link[qtl2scan]{top_snps}}}

\item{snp_col}{color of SNP vertical lines (default "grey70")}

\item{extend}{extend region for SNPs in bp (default 5000)}

\item{...}{additional arguments (not used)}
}
\value{
data frame of gene information (invisible)
}
\description{
Plot genes as rectangles followed by names.
Stagger genes for easy reading.
Written original by Dan Gatti 2013-02-13
}
\references{
\url{https://github.com/dmgatti/DOQTL/blob/master/R/gene.plot.R}
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
   Daniel Gatti, \email{Dan.Gatti@jax.org}
}
\keyword{hplot}
