% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_exon.R
\name{plot_gene_exon}
\alias{plot_gene_exon}
\alias{autoplot.gene_exon}
\alias{plot.gene_exon}
\title{Plot of exons for a gene with SNPs}
\usage{
plot_gene_exon(gene_exon, top_snps_tbl = NULL, plot_now = TRUE,
  genes = unique(gene_exon$gene), ...)

\method{autoplot}{gene_exon}(x, ...)

\method{plot}{gene_exon}(x, ...)
}
\arguments{
\item{top_snps_tbl}{table from \code{\link[qtl2scan]{top_snps}}}

\item{plot_now}{plot now if TRUE}

\item{...}{arguments passed along to \code{\link{gene_plot}}}

\item{exon_tbl}{tbl of feature information from \code{\link{get_mgi_features}}}
}
\value{
list of ggplots (see \code{\link{gene_plot}})
}
\description{
Uses \code{\link{gene_plot}} to plot genes, exons, mRNA with SNPs.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{hplot}
