% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mediate1.R
\name{plot_mediate1}
\alias{plot_mediate1}
\title{Plot mediate1 object}
\usage{
plot_mediate1(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link[qtl2pattern]{allele1}}.}

\item{...}{Other parameters ignored.}
}
\description{
Plot mediation by expression phenotypes in region.
}
