% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scan1.R
\name{plot_scan1}
\alias{plot.scan1}
\alias{plot_scan1}
\title{Plot a genome scan}
\usage{
plot_scan1(x, lodcolumn = 1, chr = NULL, add = FALSE, gap = 25,
  bgcolor = "gray90", altbgcolor = "gray85", ...)

\method{plot}{scan1}(x, lodcolumn = 1, chr = NULL, add = FALSE,
  gap = 25, bgcolor = "gray90", altbgcolor = "gray85", ...)
}
\arguments{
\item{x}{Output of \code{\link[qtl2scan]{scan1}}.}

\item{lodcolumn}{LOD score column to plot (a numeric index, or a
character string for a column name). Only one value allowed.}

\item{chr}{Selected chromosomes to plot; a vector of character
strings.}

\item{add}{If TRUE, add to current plot (must have same map and
chromosomes).}

\item{gap}{Gap between chromosomes.}

\item{bgcolor}{Background color for the plot.}

\item{altbgcolor}{Background color for alternate chromosomes.}

\item{...}{Additional graphics parameters.}
}
\value{
None.
}
\description{
Plot LOD curves for a genome scan
}
\examples{
# load qtl2geno package for data and genoprob calculation
library(qtl2geno)

# read data
iron <- read_cross2(system.file("extdata", "iron.zip", package="qtl2geno"))

# calculate genotype probabilities
probs <- calc_genoprob(iron, step=1, error_prob=0.002)

# grab phenotypes and covariates; ensure that covariates have names attribute
pheno <- iron$pheno
covar <- match(iron$covar$sex, c("f", "m")) # make numeric
names(covar) <- rownames(iron$covar)
Xcovar <- get_x_covar(iron)

# perform genome scan
library(qtl2scan)
out <- scan1(probs, pheno, addcovar=covar, Xcovar=Xcovar)

# plot the results for selected chromosomes
ylim <- c(0, maxlod(out)*1.02) # need to strip class to get overall max LOD
chr <- c(2,7,8,9,15,16)
plot(out, lodcolumn=1:2, chr=chr, ylim=ylim, col=c("darkslateblue","violetred"), 
     legend.position=c(0.1,0.9))

# plot just one chromosome
plot(out, chr=8, lodcolumn=1:2, ylim=ylim, col=c("darkblue","violetred"))

# lodcolumn can also be a column name
plot(out, chr=8, lodcolumn=c("liver","spleen"), ylim=ylim, col=c("darkblue","violetred"))
}
\seealso{
\code{\link{plot_coef}}, \code{\link{plot_coefCC}}, \code{\link{plot_snpasso}}
}

