% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scan_pattern.R
\name{plot_scan_pattern}
\alias{plot_scan_pattern}
\alias{autoplot.scan_pattern}
\title{Plot scan pattern}
\usage{
plot_scan_pattern(x, plot_type = c("lod", "coef", "coef_and_lod"),
  patterns = x$patterns$founders, columns = 1:3, min_lod = 3,
  ylim_coef = c(-2, 2), ...)

\method{autoplot}{scan_pattern}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{scan_pattern}}}

\item{plot_type}{type of plot from \code{c("lod","coef")}}

\item{patterns}{allele patterns to plot (default all)}

\item{columns}{columns used for coef plot}

\item{min_lod}{minimum LOD peak for contrast to be retained}

\item{ylim_coef}{vertical limits for coef plot}

\item{...}{additional parameters}
}
\description{
Plot scan pattern
}
