% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_to_pattern.R
\name{sdp_to_pattern}
\alias{sdp_to_pattern}
\title{Convert sdp to pattern}
\usage{
sdp_to_pattern(sdp, haplos, symmetric = TRUE)
}
\arguments{
\item{sdp}{vector of sdp values}

\item{haplos}{letter codes for haplotypes (required)}

\item{symmetric}{make patterns symmetric if \code{TRUE}}
}
\value{
vector of letter patterns
}
\description{
Convert strain distribution pattern (sdp) to letter pattern.
Taken from package `qtl2pattern` for internal use here.
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
