% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_scan1.R
\name{summary_scan1}
\alias{summary.scan1}
\alias{summary_scan1}
\title{Summary of topsnp_pattern object}
\usage{
summary_scan1(object, type = c("best", "common"), drop = 1.5,
  show_all_snps = FALSE, ...)

\method{summary}{scan1}(object, type, ...)
}
\arguments{
\item{object}{of class \code{topsnp_pattern}}

\item{type}{type of summary}

\item{drop}{LOD drop from maximum}

\item{show_all_snps}{show all SNPs if \code{TRUE}}

\item{...}{other arguments not used}
}
\value{
table summary
}
\description{
Summary of topsnp_pattern object
}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}

