% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_snps_all.R
\name{top_snps_all}
\alias{top_snps_all}
\title{Top SNPs for all phenotypes scanned}
\usage{
top_snps_all(scan1_output, drop = 1.5, show_all_snps = TRUE)
}
\arguments{
\item{scan1_output}{output of linear mixed model for \code{phename} (see \code{\link[qtl2scan]{scan1}})}

\item{drop}{include all SNPs within \code{drop} of max LOD (default 1.5)}

\item{show_all_snps}{show all SNPs if \code{TRUE}}
}
\value{
table of top_snps at maximum lod for \code{pattern}
}
\description{
Separate fine mapping scans by allele pattern.
}
