% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_patterns.R
\name{color_patterns_pheno}
\alias{color_patterns_pheno}
\title{Set up col, pattern, shape and group for plotting.}
\usage{
color_patterns_pheno(
  scan1ggdata,
  lod,
  pattern,
  col,
  shape,
  patterns,
  facet = NULL
)
}
\arguments{
\item{scan1ggdata}{data frame to be used for plotting}

\item{lod}{matrix of LOD scores by position and pheno}

\item{pattern}{allele pattern of form \code{AB:CDEFGH}}

\item{col}{Color for \code{color} column in \code{scan1ggdata}}

\item{shape}{Shape for \code{shape} column in \code{scan1ggdata}}

\item{patterns}{Connect SDP patterns: one of \code{c("none","all","hilit")}}

\item{facet}{use \code{\link[ggplot2]{facet_wrap}} if not \code{NULL}}
}
\value{
data frame \code{scan1ggdata} with additional objects.
}
\description{
Set up col, pattern, shape and group for plotting.
}
