% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_genes.R
\name{ggplot_genes}
\alias{ggplot_genes}
\title{Plot genes}
\usage{
ggplot_genes(start, end, strand, rect_top, rect_bottom, colors, space, y,
  dir_symbol, name, xlim, xlab = "Position (Mbp)", ylab = "",
  bgcolor = "gray92", xat = NULL, legend.position = "none",
  vlines = NULL, vlines.col = "white", vlines.lwd = 1, vlines.lty = 1)
}
\arguments{
\item{start, end, strand, rect_top, rect_bottom, colors, space, y, dir_symbol, name, xlim}{usual parameters}

\item{xlab, ylab, bgcolor, xat, legend.position, vlines, vlines.col, vlines.lwd, vlines.lty, xlab, ylab, bgcolor, xat}{hidden parameters}

\item{...}{Additional graphics parameters.}
}
\description{
Plot genes at positions
}
