% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_scan1.R
\name{ggplot_scan1}
\alias{ggplot_scan1}
\title{Plot a genome scan}
\usage{
ggplot_scan1(map, lod, gap = 25, col = NULL, shape = NULL,
  pattern = NULL, facet = NULL, patterns = c("none", "all", "hilit"), ...)
}
\arguments{
\item{map}{Map of pseudomarker locations.}

\item{lod}{Matrix of lod (or other) values.}

\item{gap}{Gap between chromosomes.}

\item{col}{colors for points or lines, with labels.}

\item{pattern}{Use to group values for plotting (default = \code{NULL}); typically provided by \code{\link{plot_snpasso}} internal routine.}

\item{facet}{Plot facets if multiple phenotypes and group provided (default = \code{NULL}).}

\item{patterns}{Connect SDP patterns: one of \code{c("none","all","hilit")}.}

\item{...}{Additional graphics parameters.}

\item{bgcolor}{Background color for the plot.}

\item{altbgcolor}{Background color for alternate chromosomes.}

\item{lwd, pch, cex, col, xlim, ylim, xaxt, yaxt}{base plot parameters (coverted to ggplot use)}

\item{point_fill}{fill color for \code{\link[ggplot2]{geom_point}}}

\item{palette}{color palette for points and lines}

\item{xlab, ylab, main}{Titles for x,y and plot.}

\item{hlines, vlines}{Horizontal and vertical lines.}

\item{legend.position, legend.title}{Legend theme setting.}

\item{lines, points}{Include lines and/or points.}

\item{scales}{One of \code{c("free_x","free")}.}
}
\description{
Plot LOD curves for a genome scan
}
