% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdp_to_pattern.R
\name{sdp_to_pattern}
\alias{sdp_to_pattern}
\alias{sdp_to_logical}
\title{Convert sdp to pattern}
\usage{
sdp_to_pattern(sdp, haplos = LETTERS[1:8])

sdp_to_logical(sdp, haplos = LETTERS[1:8])
}
\arguments{
\item{sdp}{vector of sdp values}

\item{haplos}{letter codes for haplotypes (defaults to first 8 \code{LETTERS})}
}
\value{
vector of letter patterns
}
\description{
Convert strain distribution pattern (sdp) to letter pattern.
}
\examples{
\dontrun{sdp_to_pattern(snp)}

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
