% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_scan1.R
\name{summary_scan1}
\alias{summary_scan1}
\alias{summary.scan1}
\title{Summary of scan1 object}
\usage{
summary_scan1(object, lodcolumn = seq_len(ncol(object$lod)),
  chr = names(object$map), type = c("common", "best"), drop = 1.5,
  show_all_snps = TRUE, ...)

\method{summary}{scan1}(object, ...)
}
\arguments{
\item{object}{object from \code{\link[qtl2scan]{scan1}}}

\item{lodcolumn}{one or more lod columns}

\item{chr}{one or more chromosome IDs}

\item{type}{type of summary}

\item{drop}{LOD drop from maximum}

\item{show_all_snps}{show all SNPs if \code{TRUE}}

\item{...}{other arguments not used}
}
\value{
tbl summary
}
\description{
Summary of scan1 object
}
\examples{
dontrun(summary(object))

}
\author{
Brian S Yandell, \email{brian.yandell@wisc.edu}
}
\keyword{utilities}
